Squashed commit of the following:
commit a3d91a3522d1256cf63e08027cd5505e5736406f
Author: arark <knkr5456@gmail.com>
Date:   Fri Oct 22 10:38:28 2021 +0900

    Test(firefuse)/rearrange tests (#8)

    * test: split test cases into correspond files

    * fix(Extends, Exact): return correct ansewer using tuple

    * chore(tsconfig): turn on strictNullCheck

commit e37f3faae6947e5857287deeb410fe52c0bd7b74
Merge: 47ccfb1 7aedcb8
Author: arark <knkr5456@gmail.com>
Date:   Thu Oct 21 19:22:32 2021 +0900

    Merge pull request #7 from Hagihara-A/feat(firefuse)/enable-exactOptionalPropertyTypes

    Feat(firefuse)/enable exact optional property types

commit 7aedcb82ecbe51673935fe5bb23ffc995bb9b00a
Merge: 4a9184f 47ccfb1
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 18:44:49 2021 +0900

    Merge branch 'master' into feat(firefuse)/enable-exactOptionalPropertyTypes

commit 4a9184f2334a8f0a7efceb8dd72b83a09909106a
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 18:41:37 2021 +0900

    refactor(Extends): switch Extends<E, A>to Extends<A, E>

commit ef0c8187403f44eab509a5c149efe79d84591129
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 18:37:12 2021 +0900

    fix(doc): export Doc interface

commit 68b28e27df128c7d32db9b869f8fba0e653d2fdc
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 18:36:53 2021 +0900

    test(where): test of LegalValue and LegalOperation

commit 58ca688a37ff8d8d4a3ef128fd675c6109649eb9
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 18:36:30 2021 +0900

    fix(LegalValue): exclude undfined in optional property

commit a3860c839b070f7535d8918b1056b61246edcbcf
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 18:28:21 2021 +0900

    refactor: move OR, OverWrite, Defined into proper file

commit 43c4295181b49f562b797404b07f8358f0131eb4
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 13:27:51 2021 +0900

    Squashed commit of the following:

    commit 47ccfb157be337cc7fb55cd2625bb26d85bda93c
    Merge: 65e6f9a fc74517
    Author: arark <knkr5456@gmail.com>
    Date:   Thu Oct 21 13:25:30 2021 +0900

        Merge pull request #6 from Hagihara-A/refactor(firefuse)/arrange-files

        chore: add `.js` extension for ESM

    commit fc74517c50d6c51b781d3ed7a90e11a6f32f713c
    Author: Arark <knkr5456@gmail.com>
    Date:   Thu Oct 21 13:22:35 2021 +0900

        chore(jest): setup ESM testing

    commit e860313c06035905b18b91820b108817f1b46b8b
    Author: Arark <knkr5456@gmail.com>
    Date:   Thu Oct 21 13:22:18 2021 +0900

        chore(tests): add `.js` in test imports

    commit cc865db793e42fbf9cfbc044395f869ef84456a6
    Author: Arark <knkr5456@gmail.com>
    Date:   Thu Oct 21 13:11:56 2021 +0900

        chore: add `.js` extension for ESM

commit 47ccfb157be337cc7fb55cd2625bb26d85bda93c
Merge: 65e6f9a fc74517
Author: arark <knkr5456@gmail.com>
Date:   Thu Oct 21 13:25:30 2021 +0900

    Merge pull request #6 from Hagihara-A/refactor(firefuse)/arrange-files

    chore: add `.js` extension for ESM

commit fc74517c50d6c51b781d3ed7a90e11a6f32f713c
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 13:22:35 2021 +0900

    chore(jest): setup ESM testing

commit e860313c06035905b18b91820b108817f1b46b8b
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 13:22:18 2021 +0900

    chore(tests): add `.js` in test imports

commit cc865db793e42fbf9cfbc044395f869ef84456a6
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 13:11:56 2021 +0900

    chore: add `.js` extension for ESM

commit 23ae2c447a2d7f9e25af9d1e8c6bfae2502609bf
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 13:07:38 2021 +0900

    chore(tsconfig): turn on `exactOptionalPropertyTypes`

commit 65e6f9afec7ff0ddfe174d33f4e1b0014ec2b175
Merge: 3733116 bd92f8b
Author: arark <knkr5456@gmail.com>
Date:   Thu Oct 21 07:36:58 2021 +0900

    Merge pull request #5 from Hagihara-A/refactor(firefuse)/arrange-files

    refactor: split components into files

commit bd92f8b906aaf23a30897209dcc89ca8d32d26f0
Author: Arark <knkr5456@gmail.com>
Date:   Thu Oct 21 07:24:21 2021 +0900

    refactor: split components into files
