+++ firefuse/src/constraint/orderby.ts
@@ -1,0 +1,15 @@
+import * as firestore from "firebase/firestore";
+
+import { DocumentData, KeyofPrimitive } from "../index.js";
+
+export const orderBy = <T extends DocumentData>() => {
+  return <F extends KeyofPrimitive<T> & string>(
+    field: F,
+    order?: firestore.OrderByDirection
+  ) => firestore.orderBy(field, order) as OrderByConstraint<F>;
+};
+export interface OrderByConstraint<F extends string>
+  extends firestore.QueryConstraint {
+  readonly type: Extract<firestore.QueryConstraintType, "orderBy">;
+  field: F;
+}
